refactor(test): add utils file for test
